+++ bin/check-es-file-descriptors.rb
@@ -67,7 +67,7 @@ class ESFileDescriptors < Sensu::Plugin::Check::CLI
          default: 80
 
   def get_es_resource(resource)
-    r = RestClient::Resource.new("http://#{config[:host]}:#{config[:port]}/#{resource}", timeout: config[:timeout])
+    r = RestClient::Resource.new("http://#{config[:host]}:#{config[:port]}#{resource}", timeout: config[:timeout])
     JSON.parse(r.get)
   rescue Errno::ECONNREFUSED
     warning 'Connection refused'