+++ CHANGELOG.md
@@ -17,3 +17,14 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 
 ### Changed
 - removed cruft from /lib
+
+## [0.1.0] - 2015-07-06
+
+### Added
+- `check-es-node-status` node status check
+
+### Fixed
+- uri resource path for `get_es_resource` method
+
+### Changed
+- `get_es_resource` URI path needs to start with `/`